Inserito il context per il db

+++ Vidly/Models/Customer.cs
@@ -14,7 +14,7 @@ namespace Vidly.Models
         public string Name { get; set; }
         public bool IsSubscribedToNewsletter { get; set; }
         public MembershipType MembershipType { get; set; }
-        public byte MembershipTypeId { get; set; }
+        public int MembershipTypeId { get; set; }
 
     }
 }
